% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cb.R
\docType{package}
\name{txBenefit}
\alias{txBenefit}
\alias{txBenefit-package}
\title{txBenefit: A package for calculations pertaining to Concentration of Benefit (Cb) index.}
\description{
The txBenefit  package provides three categories of important functions:
Cb.simple(), Cb.logistic, Cb.Poisson, and cb.coxph
}
